Adding by location method, swapping tests to match, fixing bad test
